Capture Xorg crash messages in ~/.xsession-errors.
Traditionally, the xserver will dump crash messages at startup only to
stderr/stdout. Since those are lost on the console, we're having a hard
time to debug occasions where the xserver doesn't start at all, for instance
if people pass invalid xserver startup parameters on the command line.

We redirect the stdio/stderr to ~/.xsession-errors. This file is already
used for all the session programs to dump stdio/stderr messages, and
is a reasonable location to dump error messages that Xorg cannot itself
log to Xorg.0.log.

Since now xserver.c truncates an existing .xsession-errors file, we change
user.c to append stdio/stderr output to this file, so that we only truncate
the file once per startup.