Logging: Convert to using syslog().
The code is rather complex and cleaning this up to reduce open fd's
and still keep logging properly enabled and going to the right place
is just way too much work. Instead, we can just rely on syslog()
instead, which solves a lot of problems for us.

We syslog() each message, as daemon for the root parts, as user for
the rest. Each thread can just syslog whenever it wants.

An added benefit is that we lose 2 log files in /var/log, and the
output of uxlaunch is now coherently integrated with the other
daemons outputting information as uxlaunch/X/desktop starts up.