Consolekit: do this in the root stage instead.
We should be passing is-local=true probably, so let's do that in any
case to make sure that if CK is reusing this, it's set correct.

More importantly, we should not OpenSessionWithParameters as user
which is problematic no matter what. Instead, right after PAM
is a good time to setup the CK session.

This has implications for session-only use, but we're assuming that
those already have a proper CK session through some other mechanism.