v0.65 changes.
These were changes I made to uxlaunch after we stopped working on
MeeGo. I took out most of the hardcoded stuff and did a bunch of
cleanups. Now, 5+ years later, I don't have a git tree of the
individual changes but I do have a release tarball I saved, and this
is what's in them. For posterity, and, because these changes all
make sense, I don't want to lose or redo them.